Moved compatibility test to top level
